requirements.txt
Django==4.1.2
social-auth-app-django
pyyaml
djangorestframework
markdown
django-filter